app/requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2.2
django-environ==0.12.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.7.1
flake8==7.3.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
pillow==11.3.0
pycodestyle==2.14.0
pyflakes==3.4.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.2.0
